[do-not-review] Move VariableInfo into its own file to avoid circular dependency
[ghstack-poisoned]
